fix: 기존 UserSessionRepository의 회원가입 테스트를 SignUpRepository로 이전
